refactor: ssrRequest를 Request_TEMP로 리네임
